python/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "contextpack"
version = "0.1.0"
description = "ContextPack Python SDK — build, write, and encode AI context cards"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "paho-mqtt>=2.0",
    "pymodbus>=3.5",
    "httpx>=0.25",
    "cryptography>=41.0",
    "click>=8.1",
    "tiktoken>=0.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.1",
    "jsonschema>=4.20",
]

[project.scripts]
contextpack = "cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["contextpack*", "cli*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
